[dagit] State refactor for execution tabs
Summary: Refactor state management for `ExecutionTabs`, including conversion to hooks. In a followup, I'll be adding a warning dialog for when a tab is going to be removed on the Playground page.

Test Plan: Create and modify tabs in Playground, including config. Navigate away, refresh page, etc, verify that all state is maintained correctly.

Reviewers: bengotow, alangenfeld, prha, sashank

Reviewed By: bengotow, sashank

Differential Revision: https://dagster.phacility.com/D4747