add summary table for asset keys
Summary:
This is separate from the asset key structure, which we can maintain using the same schema.

Will start reading/writing to the new summary table.

Users will need to run a migration script in order for all historical asset keys to be visible in the index page.

Test Plan: back compat test

Reviewers: max, alangenfeld, schrockn

Reviewed By: max, alangenfeld

Differential Revision: https://dagster.phacility.com/D4764