(config-scaffold 2/n) Add scaffold config button to Dagit playground
Summary:
This diff adds a "scaffold missing config" button to Dagit, which is inspired by the scaffold CLI command. It does the scaffolding client side based on the config schema.

There are a few choices here that are debatable. I've highlighted them in the comments. Open to feedback about how to handle them.

Resolves: https://github.com/dagster-io/dagster/issues/3118

Test Plan:
Created solid and resource with every possible config type. Checked to make sure scaffolded config was correct.

Video:

{F351278}

Differential Revision: https://dagster.phacility.com/D4808