Handle non-existant files better during gRPC startup 1;95;0c
Summary: Right now if you pass in a file that just doesn't exist and we try to spin up a gRPC server, it will get caught in a click param error, which we don't have a great way to detect and catch - so the gRPC server startup just times out instead of raising a useful error. Instead, check this lower in the stack, when you try to actually load the file.

Test Plan: New BK coverage, load a workspace manually with a non-existant file, see a clearer error instead of a weird click param failure and then a timeout

Reviewers: alangenfeld, max, sashank

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4907