[sensors-3] add sensor cli, stored state
Summary:
Adds a generic jobs and job ticks storage table, which is structured the same as the schedules / schedule tick tables.

Currently the schedules are still only being written to the schedules table.  The plan is to have both schedules and sensors written to a unified jobs table, and to then drop the schedules only tables.

Test Plan: bk

Reviewers: schrockn, dgibson, sandyryza

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D5020