Introduce new External{Repository/Pipeline/Schedule}Origin class that uniquely identifies an external object
Summary:
This is a backwards compatible subset of https://dagster.phacility.com/D4941 that just adds the new origin type without fully removing the old one yet, particularly in the scheduler. We can remove it closer to the 0.10.0 date when we are comfortable making a breaking change.

Now that we are going to be persisting pipeline origins in the run database forever, it's worth making sure that what we persist actually reflects what we want to store forever.

Right now we are using origins for two different purposes:
- PipelinePythonOrigin is used in user processes to identify the code pointer needed to execute a pipeline
- RepositoryOrigins are used in host processes to recreate an ExternalRepository/Pipeline/Schedule

The things that make the first thing possible (including the code pointer in the origin rather than the repository name) means that changing the code pointer breaks all of your schedule origins in the database, even if the underlying repository doesn't change.

This diff adds a new origin type for use in the later case, for use in the run queueing product.

Test Plan: BK+Azure

Reviewers: alangenfeld, schrockn, sashank, prha

Reviewed By: alangenfeld

Subscribers: johann

Differential Revision: https://dagster.phacility.com/D4994