[run-queue-7] Switch Dagit to use submit_run
Summary: The first changeover of a launch_run callsite. This should be safe to land, because with the LaunchImmediateRunCoordinator as default, changing the call site just adds a layer of indirection.

Test Plan:
Not familiar with grapqhql or dagit tests, could use some pointers here. I assume there is an existing test for `launch_pipeline_execution` in graphql but I haven't found it.

Manually: verified that both Coordinators can be used to launch runs from Dagit

Reviewers: alangenfeld, dgibson

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4919