Correctly test fall DST transition
Summary: Realized (after living through it in 2020) that I had the times wrong here in this test - in the spring, 2:30 AM doesn't exist, but it would be 1:30AM that gets repeated in the fall, not 2:30AM. So to test this correctly we need to change the time. Stiill seems to work fine.

Test Plan: BK

Reviewers: sashank, schrockn

Reviewed By: sashank

Differential Revision: https://dagster.phacility.com/D4986