[dagit] Surface repo reload errors
Summary:
Display an error in the Dagit left nav when a repo location fails to reload. Clicking on the error will display a dialog with information about the error.

When the error is resolved, reloading the repo will result in a successful state.

We should use a similar UI when initially loading a repo with errors.

Test Plan:
`RepositoryPicker.test`

Run Dagit with multiple repo locations. Add a syntax error to one, then reload that repo in Dagit. Verify that the error appears and that I can see the details when clicking the error. Remove the syntax error, try again, verify successful reload.

Reviewers: dgibson, alangenfeld

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D5021