Periodically reload pending and recent runs in the partition matrix, unify filters
Summary:
This diff makes a few more incremental improvements to the Partition matrix UI:

- The partition matrix now loads all the partition names in the current "page" before loading any run data. This makes it much faster to paginate back in time and allows us to fetch runs in each partition using the runs query, not the partitions>runs query, unifying the GraphQL for fetching and then refreshing runs.

- The partition matrix loads runs using the dagster/partition_set and dagster/partition tag filters, and runs several queries simultaneously for a 2.1x speed improvement.

- Once the partition matrix has loaded all the partitions, it starts refreshing the UI every 10 seconds by:

  + Asking for all pending runs with the dagster/partition_set tag and upserting new data into the previously fetched data
  + Asking for the last 10 runs with the dagster/partition_set tag, which allows it to detect runs being queued incrementally.

Previously, the partition matrix UI fetched everything and then the "filter" UI on the page was implemented client side. This diff switches this behavior so that applying a backfill tag, etc. alters the actual queries. This means loading less data and will also give us more options in the future to change the loading + refresh behavior when the user has scoped the UI down to a small number of runs or a  backfill.

Test Plan: Run tests

Reviewers: dish, prha, schrockn, sandyryza

Reviewed By: dish

Subscribers: alangenfeld

Differential Revision: https://dagster.phacility.com/D5018