[dagit] Downgrade graphql-playground-react
Summary:
Looks like the `graphql-playground-react` upgrade involved moving around some CSS files that our build scripts no longer know where to find, breaking the build.

Downgrade back to the previous version to clean this up. This was included in the upgrades to try to clean up some incorrect peer deps, but it's not a big deal. I'll look into it later.

Test Plan: Run `yarn build-for-python` locally, verify that there are no errors.

Reviewers: dgibson

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D5134